% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-eventset.R
\name{.create_event_set_table}
\alias{.create_event_set_table}
\title{Internal function to create contingency table for event set (full database comparison)}
\usage{
.create_event_set_table(
  object,
  full,
  event_set,
  event_type,
  object_ids,
  full_ids,
  n1,
  n_total,
  ...
)
}
\description{
Internal function to create contingency table for event set (full database comparison)
}
\keyword{internal}
