% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-eventset.R
\name{.create_event_set_table_comparison}
\alias{.create_event_set_table_comparison}
\title{Internal function to create contingency table for event set (drug comparison)}
\usage{
.create_event_set_table_comparison(
  object1,
  object2,
  event_set,
  event_type,
  object1_ids,
  object2_ids,
  n1,
  n0,
  ...
)
}
\description{
Internal function to create contingency table for event set (drug comparison)
}
\keyword{internal}
