% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-eventset.R
\name{.identify_event_set_patients}
\alias{.identify_event_set_patients}
\title{Internal function to identify patients with events in the specified set}
\usage{
.identify_event_set_patients(object, event_set, event_type, ...)
}
\description{
Internal function to identify patients with events in the specified set
}
\keyword{internal}
