% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FAERS.R
\docType{class}
\name{FAERS-class}
\alias{FAERS-class}
\alias{FAERS}
\alias{FAERSascii-class}
\alias{FAERSascii}
\alias{FAERSxml-class}
\alias{FAERSxml}
\alias{show,FAERS-method}
\alias{show,FAERSascii-method}
\alias{faers_header}
\alias{faers_header,FAERSxml-method}
\alias{faers_data}
\alias{faers_data,FAERS-method}
\alias{faers_year}
\alias{faers_year,FAERS-method}
\alias{faers_quarter}
\alias{faers_quarter,FAERS-method}
\alias{faers_period}
\alias{faers_period,FAERS-method}
\alias{faers_deleted_cases}
\alias{faers_deleted_cases,FAERSascii-method}
\alias{faers_field}
\title{FAERS class}
\usage{
\S4method{show}{FAERS}(object)

\S4method{show}{FAERSascii}(object)

faers_header(object)

\S4method{faers_header}{FAERSxml}(object)

faers_data(object, ...)

\S4method{faers_data}{FAERS}(object)

faers_year(object)

\S4method{faers_year}{FAERS}(object)

faers_quarter(object)

\S4method{faers_quarter}{FAERS}(object)

faers_period(object)

\S4method{faers_period}{FAERS}(object)

faers_deleted_cases(object, ...)

\S4method{faers_deleted_cases}{FAERSascii}(object)
}
\arguments{
\item{object}{A \link{FAERS} object.}

\item{...}{Other arguments passed to specific methods.}
}
\description{
Provide a container for FAERS Quarterly Data file
}
\details{
\itemize{
\item \code{faers_data}: Extract the \code{data} slot.
\item \code{faers_year}: Extract the \code{year} slot.
\item \code{faers_quarter}: Extract the \code{quarter} slot.
\item \code{faers_period}: Extract the \code{period} slot (just Concatenate the year and
quarter slot).
\item \code{faers_deleted_cases}: Extract the \code{deletedCases} slot.
}
}
\section{Slots}{

\describe{
\item{\code{year}}{An integer specifies the year information.}

\item{\code{quarter}}{A string specifies the quarter information.}

\item{\code{data}}{For \code{FAERSxml}, a \link[data.table:data.table]{data.table}. For
\code{FAERSascii}, a list of \link[data.table:data.table]{data.table}.}

\item{\code{meddra}}{A \link[data.table:data.table]{data.table} or \code{NULL} representing
the meddra data used for standardization.}

\item{\code{format}}{A string of "ascii" or "xml" indicates the file format used.}

\item{\code{deletedCases}}{A list of integers, as of 2019 Quarter one there are new
files that lists deleted cases.}
}}

