% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{FAERS-extractor}
\alias{FAERS-extractor}
\alias{faers_year}
\alias{faers_year,FAERS-method}
\alias{faers_year,ListOfFAERS-method}
\alias{faers_quarter}
\alias{faers_quarter,FAERS-method}
\alias{faers_quarter,ListOfFAERS-method}
\alias{faers_period}
\alias{faers_period,FAERS-method}
\alias{faers_data}
\alias{faers_data,FAERS-method}
\alias{faers_data,ListOfFAERS-method}
\alias{faers_field}
\alias{faers_field,FAERS-method}
\alias{faers_field,ListOfFAERS-method}
\alias{faers_fields}
\alias{faers_fields,FAERS-method}
\title{Extract data slot of a \link{FAERS} object}
\usage{
faers_year(object)

\S4method{faers_year}{FAERS}(object)

\S4method{faers_year}{ListOfFAERS}(object)

faers_quarter(object)

\S4method{faers_quarter}{FAERS}(object)

\S4method{faers_quarter}{ListOfFAERS}(object)

faers_period(object)

\S4method{faers_period}{FAERS}(object)

\S4method{faers_period}{FAERS}(object)

faers_data(object, ...)

\S4method{faers_data}{FAERS}(object)

\S4method{faers_data}{ListOfFAERS}(object, year = NULL, quarter = NULL, period = NULL)

faers_field(object, ...)

\S4method{faers_field}{FAERS}(object, field)

\S4method{faers_field}{ListOfFAERS}(object, field, year = NULL, quarter = NULL, period = NULL)

faers_fields(object, ...)

\S4method{faers_fields}{FAERS}(object, fields)
}
\arguments{
\item{object}{A \link{ListOfFAERS} object.}

\item{...}{Other arguments passed to specific methods.}

\item{value}{For \link{faers_year}, a numeric coerced to integer, for
\link{faers_quarter}, a string.}
}
\description{
Extract data slot of a \link{FAERS} object
}
