% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{faers_before_period}
\alias{faers_before_period}
\title{Test whether years and quarters are before specified period}
\usage{
faers_before_period(years, quarters, y, q, inclusive = TRUE)
}
\arguments{
\item{years}{An atomic integer indicates years to test.}

\item{quarters}{An atomic character indicates quarters to test, only "q1",
"q2", "q3", and "q4" are allowed.}

\item{y}{An integer, specified period year.}

\item{q}{An string, specified period quarter.}

\item{inclusive}{A bool, whether to include the period specifid.}
}
\value{
An atomic logical with the same length of the max length of \code{years}
and \code{quarters}.
}
\description{
Test whether years and quarters are before specified period
}
