% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{faers_clearcache}
\alias{faers_clearcache}
\title{Remove caches}
\usage{
faers_clearcache(caches = NULL, force = FALSE)
}
\arguments{
\item{caches}{An atomic character, indicates what caches to remove? Only
"metadata", "fdadrugs", and "athena" can be used. If \code{NULL}, all caches will be
removed.}

\item{force}{logical.  Should permissions be changed (if possible) to
    allow the file or directory to be removed?}
}
\value{
Path of the deleted directory invisiblely
}
\description{
Remove caches
}
\examples{
faers_clearcache()
}
