% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{faers_combine}
\alias{faers_combine}
\title{Combine a list FAERS objects into one}
\usage{
faers_combine(x)
}
\arguments{
\item{x}{A list of \link{FAERSxml} or \link{FAERSascii} objects.}
}
\value{
A \link{FAERSxml} or \link{FAERSascii} object.
}
\description{
Packed all \link{FAERSascii} or \link{FAERSxml} objects into a single \link{FAERSascii} or
\link{FAERSxml} object
}
