% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{faers_load}
\alias{faers_load}
\title{Load data attached in faers package}
\usage{
faers_load(nm)
}
\arguments{
\item{nm}{A string of the data name.
Available name: "irAEs".}
}
\value{
\itemize{
\item \code{irAEs}: A \link[data.table:data.table]{data.table}
}
}
\description{
Load data attached in faers package
}
\details{
\itemize{
\item \code{irAEs}: Immune-related adverse events examined in ICI-associated adverse
events
}
}
\examples{
faers_load("irAEs")
}
\references{
\itemize{
\item Chen Chen, Bin Wu, ChenYu Zhang, Ting Xu, Immune-related adverse events
associated with immune checkpoint inhibitors: An updated comprehensive
disproportionality analysis of the FDA adverse event reporting system,
International Immunopharmacology
}
}
