% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{ListOfFAERS}
\alias{ListOfFAERS}
\alias{print.ListOfFAERS}
\title{Create ListOfFAERS class}
\usage{
ListOfFAERS(x)

\method{print}{ListOfFAERS}(x, ...)
}
\arguments{
\item{x}{A \link{ListOfFAERS} object.}

\item{...}{Not used currently.}
}
\value{
A \code{ListOfFAERS} object
}
\description{
Packed a list of \link{FAERSascii} or \link{FAERSxml} objects into a single object
}
