% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phv.R
\name{phv_signal}
\alias{phv_signal}
\alias{phv_ror}
\alias{phv_prr}
\alias{phv_bcpnn_norm}
\alias{phv_bcpnn_mcmc}
\title{Pharmacovigilance Analysis}
\usage{
phv_ror(a, b, c, d, alpha = 0.05)

phv_prr(a, b, c, d, alpha = 0.05)

phv_bcpnn_norm(a, b, c, d, alpha = 0.05)

phv_bcpnn_mcmc(a, b, c, d, alpha = 0.05, n_mcmc = 100000L)
}
\arguments{
\item{a}{also referred to as \code{n11} as this is the count of event of interest
under exposure of interest.}

\item{b}{also referred to as \code{n10} as this is the count of \emph{not} event
of interest under exposure of interest.}

\item{c}{also referred to as \code{n01} as this is the count of event of interest
under \emph{not} exposure of interest.}

\item{d}{also referred to as \code{n00} as this is the count of \emph{not} event
of interest under \emph{not} exposure of interest.}

\item{alpha}{Level of significance, for construction of the confidence
intervals.}

\item{n_mcmc}{number of MCMC simulations per \code{(a,b,c,d)}-tuple.}
}
\value{
A \link[data.table:data.table]{data.table}
}
\description{
In the context of pharmacovigilance studies, disproportionality analysis
primarily served as a tool to evaluate possible association between a
specific adverse event and a particular drug which can then be investigated
through clinical assessment of individual case reports. Pharmacovigilance,
also known as drug safety. Following function can be used to Signal mining
}
\section{Contingency table}{
\tabular{llll}{
    \tab ADR of interest \tab Other ADRs \tab Total \cr
   Drug of interest \tab a=n11 \tab b=n10 \tab a + b \cr
   Other drugs \tab c=n01 \tab d=n00 \tab c + d \cr
   Total \tab a+c \tab b+d \tab a+b+c+d \cr
}
}

