% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{FAERS-class}
\alias{FAERS-class}
\alias{FAERS}
\alias{FAERSascii-class}
\alias{FAERSascii}
\alias{FAERSxml-class}
\alias{FAERSxml}
\alias{show,FAERS-method}
\alias{year}
\alias{year<-}
\alias{year,FAERS-method}
\alias{year<-,FAERS-method}
\alias{quarter}
\alias{quarter<-}
\alias{quarter,FAERS-method}
\alias{quarter<-,FAERS-method}
\alias{datatable}
\alias{datatable<-}
\alias{datatable,FAERS-method}
\alias{datatable<-,FAERS-method}
\alias{header}
\alias{header<-}
\alias{header,FAERSxml-method}
\alias{header<-,FAERSxml-method}
\title{FAERS classes}
\usage{
\S4method{show}{FAERS}(object)

year(object)

year(object) <- value

\S4method{year}{FAERS}(object)

\S4method{year}{FAERS}(object) <- value

quarter(object)

quarter(object) <- value

\S4method{quarter}{FAERS}(object)

\S4method{quarter}{FAERS}(object) <- value

datatable(object)

datatable(object) <- value

\S4method{datatable}{FAERS}(object)

\S4method{datatable}{FAERS}(object) <- value

header(object)

header(object) <- value

\S4method{header}{FAERSxml}(object)

\S4method{header}{FAERSxml}(object) <- value
}
\arguments{
\item{object}{A \link{FAERS} object.}

\item{value}{For \link{year}, a numeric coerced to integer, for \link{quarter}, a
string.}
}
\description{
FAERS classes
}
\section{Slots}{

\describe{
\item{\code{year:}}{An integer specifies the year information.}

\item{\code{quarter:}}{A string specifies the quarter information.}

\item{\code{datatable:}}{For \code{FAERSxml}, a \link[data.table:data.table]{data.table}. For
\code{FAERSascii}, a list of \link[data.table:data.table]{data.table}.}
}}

