% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-utils.R
\name{faers_get}
\alias{faers_get}
\alias{faers_get,FAERSascii-method}
\alias{faers_keep}
\alias{faers_keep,FAERSascii-method}
\alias{faers_filter}
\alias{faers_filter,FAERSascii-method}
\title{Methods for FAERS class}
\usage{
faers_get(object, ...)

\S4method{faers_get}{FAERSascii}(object, field)

faers_keep(object, ...)

\S4method{faers_keep}{FAERSascii}(object, primaryid = NULL)

faers_filter(object, ...)

\S4method{faers_filter}{FAERSascii}(object, field, .fn, ...)
}
\arguments{
\item{object}{A \link{FAERS} object.}

\item{...}{Other arguments passed to specific methods. For \code{faers_filter}
Other arguments passed to \code{.fn}.}

\item{field}{A string indicates the FAERS fields data applied with \code{.fn} to
extract primaryid. If \code{NULL}, the object will be passed to \code{.fn} directly.
For string, only values "demo", "drug", "indi", "ther", "reac", "rpsr", and
"outc" can be used.}

\item{primaryid}{An atomic character or integer specifies the reports to
keep. If \code{NULL}, will do nothing.}

\item{.fn}{A function or formula.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs. Functions
created from formulas have a special class. Use \code{is_lambda()} to test for
it.

If a \strong{string}, the function is looked up in \code{globalenv()}.}
}
\description{
Utils function for \link{FAERS} class.
}
\details{
\itemize{
\item \code{faers_get}: Extract a specific field \link[data.table:data.table]{data.table}
from \link{FAERS} object.
\item \code{faers_keep}: only keep data from specified \code{primaryid}. Note: \code{year},
\code{quarter}, \code{deletedCases} will be kept as the original. So make sure you
didn't filter a whole period FAERS quarterly data, in this way, it's much
better to run \link[=FAERS]{faers}.
\item \code{faers_filter}: apply a function to extract wanted \code{primaryid}, then use
\code{faers_keep} to filter.
}
}
