% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faers-package.R
\docType{package}
\name{faers-package}
\alias{faers-package}
\alias{_PACKAGE}
\title{faers: R interface for FDA Adverse Event Reporting System}
\description{
FAERS is a database for the spontaneous reporting of adverse events and medication errors involving human drugs and therapeutic biological products. faers is the bridge between FAERS and R.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@gmail.com}

}
\keyword{internal}
