% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faers-package.R
\docType{package}
\name{faers-package}
\alias{faers}
\alias{faers-package}
\title{faers: R interface for FDA Adverse Event Reporting System}
\description{
R interface for FDA Adverse Event Reporting System (FAERS).
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
