% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faers-package.R
\docType{package}
\name{faers-package}
\alias{faers-package}
\title{faers: R interface for FDA Adverse Event Reporting System}
\description{
The FDA Adverse Event Reporting System (FAERS) is a database used for the spontaneous reporting of adverse events and medication errors related to human drugs and therapeutic biological products. faers pacakge serves as the interface between the FAERS database and R. Furthermore, faers pacakge offers a standardized approach for performing pharmacovigilance analysis.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Yunuuuu/faers}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@gmail.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

Authors:
\itemize{
  \item YuXuan Song \email{yuxuan_song2021@163.com}
  \item Jiaxing Song \email{1570851599@qq.com}
  \item Caipeng Qin \email{qincaipeng@pkuph.edu.cn}
}

}
\keyword{internal}
