% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{faers_combine}
\alias{faers_combine}
\title{Combine a list FAERS objects into one}
\usage{
faers_combine(x)
}
\arguments{
\item{x}{A list of \link{FAERSxml} or \link{FAERSascii} objects.}
}
\value{
A \link{FAERSxml} or \link{FAERSascii} object.
}
\description{
Packed all \link{FAERSascii} or \link{FAERSxml} objects into a single \link{FAERSascii} or
\link{FAERSxml} object
}
\examples{
# the files included in the package are sampled
data1 <- faers_parse(
    system.file("extdata", "aers_ascii_2004q1.zip", package = "faers"),
    compress_dir = tempdir()
)
data2 <- faers_parse(
    system.file("extdata", "faers_ascii_2017q2.zip", package = "faers"),
    compress_dir = tempdir()
)
faers_combine(list(data1, data2))
}
