% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{faers_download}
\alias{faers_download}
\title{Download FAERS data}
\usage{
faers_download(years, quarters, type = c("ascii", "xml"), dir = getwd(), ...)
}
\arguments{
\item{years}{An atomic integer indicates years for which data are required.}

\item{quarters}{An atomic character, only "q1", "q2", "q3", and "q4" are
allowed.}

\item{type}{File type to download, only "ascii" and "xml" are availabe.}

\item{dir}{The destination directory for any downloads. Defaults to
current working dir.}

\item{...}{Extra handle options passed to each request
\link[curl:handle]{new_handle}}
}
\value{
An atomic character for the path of downloaded files.
}
\description{
This function downloads the FAERS data for selected years and quarters.
}
\examples{
\dontrun{
 faers_download(year = 2018, quarter = "q4", dir = tempdir())
}
}
