% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{faers_merge}
\alias{faers_merge}
\alias{faers_merge,FAERSascii-method}
\alias{faers_merge,FAERSxml-method}
\alias{faers_merge,ANY-method}
\title{Merge all FAERS field data into one}
\usage{
faers_merge(object, ...)

\S4method{faers_merge}{FAERSascii}(object, use = NULL)

\S4method{faers_merge}{FAERSxml}(object)

\S4method{faers_merge}{ANY}(object)
}
\arguments{
\item{object}{A \link{FAERSascii} or \link{FAERSxml} object.}

\item{...}{Other arguments passed to specific methods.}

\item{use}{A character vector specifying the fields to use. If \code{NULL}, all
fields will be used. Note: You'd better only merge necessary data, otherwise
all fields will consume a lot of memory.}
}
\value{
A \link[data.table:data.table]{data.table} object.
}
\description{
Merge all FAERS field data into one
}
