% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{faers_meta}
\alias{faers_meta}
\title{List of FAERS data}
\usage{
faers_meta(force = FALSE)
}
\arguments{
\item{force}{A bool. \link{faers_meta} retrieve information about all records
metadata in the FAERS Quarterly Data Extract Files Site. If a user wants to
download the metadata again and not use the cache version add the argument
\code{force=TRUE}.}
}
\value{
A \link[data.table:data.table]{data.table} reporting years, period,
quarter, and file urls and file sizes.
}
\description{
The function lists the metadata for the FAERS databases currently
available to download.
}
\examples{
faers_meta()
}
\seealso{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
