% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{faers_meta}
\alias{faers_meta}
\title{List of FAERS data}
\usage{
faers_meta(force = FALSE, internal = FALSE)
}
\arguments{
\item{force}{A boolean value. If set to \code{TRUE}, it indicates the retrieval of
information about all records' metadata in the FAERS Quarterly Data Extract
Files Site, bypassing the cache.}

\item{internal}{A boolean value. It determines whether to use the internal
data associated with the package when no cached file is available.}
}
\value{
A \link[data.table:data.table]{data.table} reporting years, period,
quarter, and file urls and file sizes.
}
\description{
The function lists the metadata for the FAERS databases currently
available to download.
}
\examples{
faers_meta()
}
\seealso{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
