% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{faers_parse}
\alias{faers_parse}
\title{Parse FAERS Quarterly Data}
\usage{
faers_parse(
  path,
  format = NULL,
  year = NULL,
  quarter = NULL,
  compress_dir = getwd()
)
}
\arguments{
\item{path}{A string specifies the path of FAERS Quarterly Data. You can pass
the FAERS zip file directly, In this way, all files in the zip file will be
extracted in \code{compress_dir}. Or, you can also uncompressed youself, and
passed the directory contained the uncompressed files.}

\item{format}{File format to used, only "ascii" and "xml" are availabe. Default:
"ascii".}

\item{year}{Year of the FAERS Quarterly Data. Coerced into integer, if
\code{NULL}, this will be extracted from path.}

\item{quarter}{String specifies quarter of the FAERS data, if \code{NULL}, this
will be extracted from path.}

\item{compress_dir}{A string specifies the directory to extract files to. It
will be created if necessary.}
}
\value{
A \link{FAERSxml} or \link{FAERSascii} object.
}
\description{
Parse FAERS Quarterly Data
}
