% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-eventset.R
\name{faers_phv_composite}
\alias{faers_phv_composite}
\alias{faers_phv_composite,FAERSascii,ANY,FAERSascii,missing-method}
\alias{faers_phv_composite,FAERSascii,ANY,missing,FAERSascii-method}
\alias{faers_phv_signal_composite}
\alias{faers_phv_signal_composite,FAERSascii-method}
\title{Create contingency table for a specific set of adverse events}
\usage{
faers_phv_composite(.object, .event_set, ..., .full, .object2)

\S4method{faers_phv_composite}{FAERSascii,ANY,FAERSascii,missing}(
  .object,
  .event_set,
  .event_type = "pt",
  ...,
  .full,
  .object2
)

\S4method{faers_phv_composite}{FAERSascii,ANY,missing,FAERSascii}(
  .object,
  .event_set,
  .event_type = "pt",
  ...,
  .full,
  .object2
)

faers_phv_signal_composite(.object, ...)

\S4method{faers_phv_signal_composite}{FAERSascii}(
  .object,
  .event_set,
  .methods = NULL,
  ...,
  .phv_signal_params = list(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{.object}{A \link{FAERSascii} object. The unique number of \code{primaryids} from
\code{.object} will be regarded as \code{n1.}.}

\item{.event_set}{A character vector of PT terms or a logical expression that
defines the set of adverse events of interest.}

\item{...}{Other arguments passed to specific methods.}

\item{.full}{A \link{FAERSascii} object with data from full database.}

\item{.object2}{A \link{FAERSascii} object with data from another interested drug.}

\item{.methods}{Analysis methods to use (passed to \link{phv_signal}).}

\item{.phv_signal_params}{Other arguments passed to \link{phv_signal}.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}} instance
    defining the parallel back-end to be used during evaluation.
  }
}
\value{
A \link[data.table:data.table]{data.table} object with contingency tables
for the specified event set.
}
\description{
Build contingency tables for disproportionality analysis focusing on a specific
set of adverse events (e.g., irAEs). This function extends \code{faers_phv_table}
to handle composite event definitions.
}
\seealso{
\link{phv_signal}
}
