% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{faers_standardize}
\alias{faers_standardize}
\alias{faers_standardize,FAERSascii-method}
\title{Standardize FAERS Quarterly Data for Preferred Term and drug names}
\usage{
faers_standardize(object, ...)

\S4method{faers_standardize}{FAERSascii}(object, meddra_path, add_smq = TRUE)
}
\arguments{
\item{object}{A \link{FAERSascii} object.}

\item{...}{Other arguments passed to specific methods.}

\item{meddra_path}{A string, define the path of MedDRA directory.}

\item{add_smq}{A bool, indicates whether Standardised MedDRA Queries should
be added. If \code{TRUE}, "smq_content.asc", and "smq_list.asc" must exist.}
}
\value{
A \link{FAERSascii} object.
}
\description{
Standardize FAERS Quarterly Data for Preferred Term and drug names
}
\seealso{
\url{https://www.meddra.org/}
}
