% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{faers_tidy}
\alias{faers_tidy}
\alias{faers_tidy,FAERSascii-method}
\alias{faers_tidy,FAERSxml-method}
\alias{faers_tidy,ListOfFAERS-method}
\title{Tidy up faers datatable}
\usage{
faers_tidy(object, ...)

\S4method{faers_tidy}{FAERSascii}(object, infos = NULL, all = TRUE)

\S4method{faers_tidy}{FAERSxml}(object)

\S4method{faers_tidy}{ListOfFAERS}(object, ...)
}
\arguments{
\item{object}{A \link{FAERSascii} or \link{FAERSxml} object.}

\item{...}{Other arguments passed to specific methods. For \link{ListOfFAERS}
method, other arguments passed to \link{FAERSascii} or \link{FAERSxml} method.}

\item{infos}{Index (integer, logistical, character all will be okay) to
select informations included in the final object. If \code{NULL}, all infomations
will be used, Note: this will take much memory.}

\item{all}{A scalar logical. If \code{TRUE} rows from all \code{infos} will be
included, if \code{FALSE}, no matching row with other infos will be excluded.
\link[data.table:merge]{merge}.}
}
\value{
A \link[data.table:data.table]{data.table} object.
}
\description{
Tidy up faers datatable
}
