% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{faers_tidy}
\alias{faers_tidy}
\alias{faers_tidy,FAERSascii-method}
\alias{faers_tidy,ListOfFAERS-method}
\alias{faers_tidy,FAERSxml-method}
\title{Tidy up faers datatable with duplicate records removed}
\usage{
faers_tidy(object, ...)

\S4method{faers_tidy}{FAERSascii}(object, field)

\S4method{faers_tidy}{ListOfFAERS}(object, field)

\S4method{faers_tidy}{FAERSxml}(object)
}
\arguments{
\item{object}{A \link{FAERSascii} or \link{ListOfFAERS} object.}

\item{...}{Other arguments passed to specific methods.}

\item{field}{A string specifying the field to use.}
}
\value{
A \link[data.table:data.table]{data.table} object.
}
\description{
Tidy up faers datatable with duplicate records removed
}
