% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{faers_tidy}
\alias{faers_tidy}
\alias{faers_tidy,FAERSascii-method}
\alias{faers_tidy,ListOfFAERS-method}
\alias{faers_tidy,FAERSxml-method}
\alias{faers_tidy,ANY-method}
\title{Tidy up faers datatable with duplicate records removed}
\usage{
faers_tidy(object, ...)

\S4method{faers_tidy}{FAERSascii}(object, fields = NULL, remove_deleted_cases = TRUE)

\S4method{faers_tidy}{ListOfFAERS}(object, fields = NULL, remove_deleted_cases = TRUE)

\S4method{faers_tidy}{FAERSxml}(object)

\S4method{faers_tidy}{ANY}(object)
}
\arguments{
\item{object}{A \link{FAERSascii} or \link{ListOfFAERS} object.}

\item{...}{Other arguments passed to specific methods.}

\item{fields}{An atomic characters specifying the fields to use. If \code{NULL},
the de-duplicated dataset will be returned.}

\item{remove_deleted_cases}{If \code{TRUE}, will remove all
\link[=FAERS-class]{deletedCases} from the final result.}
}
\value{
A \link[data.table:data.table]{data.table} object.
}
\description{
Tidy up faers datatable with duplicate records removed
}
