% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ListOfFAERS.R
\docType{class}
\name{ListOfFAERS}
\alias{ListOfFAERS}
\alias{ListOfFAERS-class}
\alias{show,ListOfFAERS-method}
\alias{[[.ListOfFAERS}
\alias{[.ListOfFAERS}
\title{ListOfFAERS class}
\usage{
ListOfFAERS(x, type)

\S4method{show}{ListOfFAERS}(object)

\method{[[}{ListOfFAERS}(object, i)

\method{[}{ListOfFAERS}(object, i)
}
\arguments{
\item{x}{A list of \link{FAERSascii} or \link{FAERSxml} objects. Only one of \code{x} or
\code{type} should be provided. If you want to create a empty \code{ListOfFAERS}
object, just provide \code{type} argument.}

\item{type}{File type to used, only "ascii" and "xml" are availabe. Default:
"ascii".}

\item{object}{A \code{ListOfFAERS} object.}

\item{i}{Indices specifying elements to extract.}
}
\description{
Packed all \link{FAERSascii} or \link{FAERSxml} objects into a single object
}
\section{Slots}{

\describe{
\item{\code{container}}{A list as the container to accommodate all
\link{FAERSascii} or \link{FAERSxml} objects.}

\item{\code{type}}{A string of "ascii" or "xml" indicates the file type used.}
}}

\note{
For \code{ListOfFAERS} object, the type slot will determine the class of
objects in \code{container} slot, and can only be set when create \code{ListOfFAERS}
object using \link{ListOfFAERS}. You shouldn't change it manually with
\link[methods:slot]{slot<-}.
}
\seealso{
\linkS4class{ListOfFAERS}
}
